+++ busy_beaver/apps/poller/api/youtube_post.py
@@ -5,7 +5,7 @@ from flask.views import MethodView
 from sqlalchemy import desc
 
 from busy_beaver import config
-from busy_beaver.adapters.youtube import YouTubeClient
+from busy_beaver.common.wrappers.youtube import YouTubeClient
 from busy_beaver.common.decorators import admin_role_required
 from busy_beaver.clients import chipy_slack
 from busy_beaver.toolbox import make_response
