+++ busy_beaver/apps/retweeter/task.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 import logging
 
-from busy_beaver.adapters import KeyValueStoreClient
+from busy_beaver.common.wrappers import KeyValueStoreClient
 from busy_beaver.clients import chipy_slack, twitter
 from busy_beaver.config import TWITTER_USERNAME
 from busy_beaver.extensions import db, rq
