+++ busy_beaver/apps/slack_integration/oauth/workflow.py
@@ -1,6 +1,6 @@
 from datetime import time
 
-from busy_beaver.adapters import SlackClient
+from busy_beaver.common.wrappers import SlackClient
 from busy_beaver.clients import slack_oauth
 from busy_beaver.extensions import db
 from busy_beaver.models import SlackInstallation
