+++ tests/_utilities/factories/event_details.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
 import factory
-from busy_beaver.adapters.meetup import EventDetails as model
+from busy_beaver.common.wrappers.meetup import EventDetails as model
 
 
 def EventDetails(session):
