+++ tests/apps/slack_integration/api/event_subscription_test.py
@@ -1,7 +1,7 @@
 import pytest
 import responses
 
-from busy_beaver.adapters.slack import TimezoneInfo
+from busy_beaver.common.wrappers.slack import TimezoneInfo
 from busy_beaver.apps.slack_integration.oauth.oauth_flow import SlackOAuthFlow
 from busy_beaver.apps.slack_integration.event_subscription import app_home_handler
 from busy_beaver.apps.slack_integration.blocks import AppHome
